requirements.txt
kivy>=2.1.0
opencv-python>=4.5.5
numpy>=1.21.0
pyserial>=3.5
plyer>=2.0.0
pyttsx3>=2.90
pywin32>=305; platform_system == "Windows"
comtypes>=1.1.10
